moved all pages inside [lng] folder and changed all realted files path changes
