+++ pages/[lng]/events/index.js
@@ -1,7 +1,7 @@
-import Layout from "../../components/Layout";
-import { charityAPI } from "../../clients";
+import Layout from "../../../components/Layout";
+import { charityAPI } from "../../../clients";
 
-const Donations = ({
+const Events = ({
   footerData,
   ContactsData,
   logoData,
@@ -16,20 +16,21 @@ const Donations = ({
       socialMediasData={socialMediasData}
       pagesData={pagesData}
     >
-      Donations components goes here
+      Events components goes here
     </Layout>
   );
 };
 
-export  async function getServerSideProps({params:{lng}}) {
-  const { default: lngDict = {} } = await import(`../../locales/${lng}.json`);
-  
+export async function getServerSideProps({params :{lng}}) {
+  const { default: lngDict = {} } = await import(`../../../locales/${lng}.json`);
+  const getCharityAPI = charityAPI(lng);
+
   return Promise.all([
-    charityAPI("/main-contacts"),
-    charityAPI("/logo"),
-    charityAPI("/socialmedias"),
-    charityAPI("/pages"),
-    charityAPI("/footer"),
+    getCharityAPI("/main-contacts"),
+    getCharityAPI("/logo"),
+    getCharityAPI("/socialmedias"),
+    getCharityAPI("/pages"),
+    getCharityAPI("/footer"),
   ]).then(
     ([
       { data: ContactsData },
@@ -53,4 +54,4 @@ export  async function getServerSideProps({params:{lng}}) {
   );
 }
 
-export default Donations;
+export default Events;
